+++ src/com/seancheey/data/fish/RectFish.java
@@ -13,7 +13,7 @@ public class RectFish extends Fish {
 	 */
 	private static final long serialVersionUID = 1L;
 
-	public RectFish(int width, int height, double x, double y, double vx,
+	public RectFish(double width, double height, double x, double y, double vx,
 			double vy, Pond pond) {
 		super(width, height, x, y, vx, vy, pond, ImagePond
 				.get("fish - redfish"));
