+++ src/com/seancheey/data/fish/RoundFish.java
@@ -13,7 +13,7 @@ public class RoundFish extends Fish {
 	 */
 	private static final long serialVersionUID = 1L;
 
-	public RoundFish(int width, int height, double x, double y, double vx,
+	public RoundFish(double width, double height, double x, double y, double vx,
 			double vy, Pond pond) {
 		super(width, height, x, y, vx, vy, pond, ImagePond
 				.get("fish - bluefish"));
